requirements.txt
psycopg2-binary
requests
docopt
python-dotenv
pydantic
mypy==0.650
pyyaml
